feat(masthead): fix resizeObserver
